remove unused stores, reorganise folder structure
