+++ src/app/features/pokemon/pages/pokemon/pokemon.component.ts
@@ -4,7 +4,7 @@ import { MatTabsModule } from "@angular/material/tabs";
 import { ActivatedRoute, RouterModule } from "@angular/router";
 import { PokemonSummaryComponent } from "../../components/pokemon-summary/pokemon-summary.component";
 import { PokemonStatsTabComponent } from "../../components/tabs/pokemon-stats-tab/pokemon-stats-tab.component";
-import { PokemonStore } from "../../../../shared/store/pokemon.store";
+import { PokemonStore } from "../../../../shared/+state/pokemon.store";
 import { PokemonDetailsTabComponent } from "../../components/tabs/pokemon-details-tab/pokemon-details-tab.component";
 
 @Component({
