+++ src/app/shared/+state/data/version-group.constants.ts
