+++ src/app/shared/+state/features/selected-entity.feature.ts
